Revert changes on the client-side as they are not required
