+++ servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/HeaderUtils.java
@@ -19,8 +19,6 @@ import io.servicetalk.buffer.api.Buffer;
 import io.servicetalk.buffer.api.CharSequences;
 import io.servicetalk.concurrent.api.Publisher;
 import io.servicetalk.concurrent.api.ScanWithMapper;
-import io.servicetalk.concurrent.api.Single;
-import io.servicetalk.concurrent.api.SingleTerminalSignalConsumer;
 import io.servicetalk.concurrent.api.TerminalSignalConsumer;
 import io.servicetalk.concurrent.internal.CancelImmediatelySubscriber;
 import io.servicetalk.http.api.EmptyHttpHeaders;
@@ -144,12 +142,13 @@ final class HeaderUtils {
                                                      final StreamingHttpRequest request) {
         return setContentLength(request, request.messageBody(),
                 shouldAddZeroContentLength(request.method()) ? HeaderUtils::updateContentLength :
-                        HeaderUtils::updateRequestContentLengthNonZero, protocolVersion);
+                        HeaderUtils::updateRequestContentLengthNonZero, protocolVersion, /* propagateCancel */ false);
     }
 
     static Publisher<Object> setResponseContentLength(final HttpProtocolVersion protocolVersion,
                                                       final StreamingHttpResponse response) {
-        return setContentLength(response, response.messageBody(), HeaderUtils::updateContentLength, protocolVersion);
+        return setContentLength(response, response.messageBody(), HeaderUtils::updateContentLength, protocolVersion,
+                /* propagateCancel */ true);
     }
 
     private static void updateRequestContentLengthNonZero(final int contentLength, final HttpHeaders headers) {
@@ -215,7 +214,9 @@ final class HeaderUtils {
     }
 
     static Publisher<Object> flatEmptyMessage(final HttpProtocolVersion protocolVersion,
-                                              final HttpMetaData metadata, final Publisher<Object> messageBody) {
+                                              final HttpMetaData metadata,
+                                              final Publisher<Object> messageBody,
+                                              final boolean propagateCancel) {
         assert emptyMessageBody(metadata, messageBody);
         // HTTP/2 and above can write meta-data as a single frame with endStream=true flag. To check the version, use
         // HttpProtocolVersion from ConnectionInfo because HttpMetaData may have different version.
@@ -223,11 +224,15 @@ final class HeaderUtils {
                 protocolVersion.major() > 1 || !shouldAppendTrailers(protocolVersion, metadata) ? from(metadata) :
                         from(metadata, EmptyHttpHeaders.INSTANCE);
         return messageBody == empty() ? flatMessage :
-                // Subscribe to the messageBody publisher to trigger any applied transformations, but ignore its
+                // 1. Subscribe to the messageBody publisher to trigger any applied transformations, but ignore its
                 // content because the PayloadInfo indicated it's effectively empty and does not contain trailers.
-                // Because `concat` won't subscribe to the messageBody in case of cancellation or an error, we use
-                // afterFinally to guarantee messageBody sees cancel too. Otherwise, observers won't complete exchange.
-                flatMessage.afterFinally(new TerminalSignalConsumer() {
+                //
+                // 2. Because `concat` won't subscribe to the messageBody in case of cancellation or an error, we use
+                // `afterFinally` to guarantee messageBody sees cancel too. This is necessary when we have to preserve
+                // propagation of all signals through Reactive Streams chain. Otherwise, some state machines won't
+                // trigger. For example, BeforeFinallyHttpOperator on the server-side won't see any signals and
+                // observers won't complete the exchange.
+                (propagateCancel ? flatMessage.afterFinally(new TerminalSignalConsumer() {
                     @Override
                     public void onComplete() {
                         // noop, rely on `concat`
@@ -246,33 +251,7 @@ final class HeaderUtils {
                     private void cancelMessageBody() {
                         toSource(messageBody).subscribe(CancelImmediatelySubscriber.INSTANCE);
                     }
-                }).concat(messageBody.ignoreElements());
-    }
-
-    static Publisher<Object> flatMessage(final HttpMetaData metadata, final Publisher<Object> messageBody,
-                                         final boolean deferSubscribe) {
-        // Because `concat` won't subscribe to the messageBody in case of cancellation or an error, we use
-        // afterFinally to guarantee messageBody sees cancel too. Otherwise, observers won't complete exchange.
-        return Single.<Object>succeeded(metadata).afterFinally(new SingleTerminalSignalConsumer<Object>() {
-                    @Override
-                    public void onSuccess(@Nullable final Object result) {
-                        // noop, rely on `concat`
-                    }
-
-                    @Override
-                    public void onError(final Throwable throwable) {
-                        cancelMessageBody();
-                    }
-
-                    @Override
-                    public void cancel() {
-                        cancelMessageBody();
-                    }
-
-                    private void cancelMessageBody() {
-                        toSource(messageBody).subscribe(CancelImmediatelySubscriber.INSTANCE);
-                    }
-                }).concat(messageBody, deferSubscribe);
+                }) : flatMessage).concat(messageBody.ignoreElements());
     }
 
     private static final class ContentLengthList<T> extends ArrayList<T> {
@@ -298,10 +277,11 @@ final class HeaderUtils {
     private static Publisher<Object> setContentLength(final HttpMetaData metadata,
                                                       final Publisher<Object> messageBody,
                                                       final BiIntConsumer<HttpHeaders> contentLengthUpdater,
-                                                      final HttpProtocolVersion protocolVersion) {
+                                                      final HttpProtocolVersion protocolVersion,
+                                                      final boolean propagateCancel) {
         if (emptyMessageBody(metadata, messageBody)) {
             contentLengthUpdater.apply(0, metadata.headers());
-            return flatEmptyMessage(protocolVersion, metadata, messageBody);
+            return flatEmptyMessage(protocolVersion, metadata, messageBody, propagateCancel);
         }
         return messageBody.collect(() -> null, (reduction, item) -> {
             if (reduction == null) {
